Create Project-2
We generated word frequencies for the entire Harry Potter series, with and without stopwords removed. We Examine the top 10 most frequent words in the series and comment upon what you expect to see and any discrepancies from that.
We generated stopword-removed word frequencies for each individual book in the Harry Potter series.  We Present the top 10 most frequent words for each of the seven books in a ggplot bar graph.
We Choose any single Harry Potter book of the seven and obtain a word cloud of it.  The words in the word cloud should be representative of the book, i.e. stopwords should be removed.
Two other authors, Suzanne Collins (Hunger Games series) and Rick Riordan (Percy Jackson series), also have works excerpted in the corpus folder.  We chose one author and find their (stopword-removed) word frequencies for their books in the corpus.  We Presented the top10 most frequent (non-stop word) words for the author in a ggplot bar graph.
We Created a jittered scatterplot contrasting the word frequencies of your author from the word frequencies of J.K. Rowling.  
We also Calculated types and hapax measures for J.K. Rowling and your author, making sure to detail how our measures were obtained.